# sodaya-api-hexagonal/openapi/v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SodaYa API",
        "version": "1.0.0",
        "description": "API de SodaYa: men\u00fa del d\u00eda y pedidos para llevar de varias sodas."
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
...
    "paths": {
...
        "/cocina/cierres/hoy": {
            "post": {
                "operationId": "closure.closeToday",
                "tags": [
                    "Closure"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CloseForTodayRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`ClosureResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ClosureResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/InvalidData"
                    }
                }
            }
        },
        "/cocina/cierres/{cierre}": {
            "delete": {
                "operationId": "closure.destroy",
                "tags": [
                    "Closure"
                ],
                "parameters": [
                    {
                        "name": "cierre",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CloseForTodayRequest": {
                "type": "object",
                "properties": {
                    "motivo": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 200
                    }
                },
                "title": "CloseForTodayRequest"
            },
            "ClosureResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "fecha": {
                        "type": "string"
                    },
                    "motivo": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "fecha",
                    "motivo"
                ],
                "title": "ClosureResource"
            },
...
        },
        "responses": {
            "NotFound": {
                "description": "Recurso no encontrado",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "format": "uri"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "detail": {
                                    "type": "string"
                                },
                                "instance": {
                                    "type": "string",
                                    "format": "uri"
                                }
                            },
                            "required": [
                                "type",
                                "title",
                                "status",
                                "detail",
                                "instance"
                            ]
                        }
                    }
                }
            },
            "InvalidData": {
                "description": "Datos inv\u00e1lidos",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "format": "uri"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "detail": {
                                    "type": "string"
                                },
                                "instance": {
                                    "type": "string",
                                    "format": "uri"
                                },
                                "errores": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "type",
                                "title",
                                "status",
                                "detail",
                                "instance"
                            ]
                        }
                    }
                }
            }
        }
    }
}